requirements.txt
streamlit
pypdf2
pandas
pillow
pytesseract